+++ src/main/java/obro1961/chatpatches/mixin/MinecraftClientMixin.java
@@ -1,14 +1,9 @@
 package obro1961.chatpatches.mixin;
 
-import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
-import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.MinecraftClient;
-import net.minecraft.client.gui.screen.Screen;
-import obro1961.chatpatches.ChatPatches;
 import obro1961.chatpatches.chatlog.ChatLog;
-import obro1961.chatpatches.gui.ChatSearchScreen;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -38,15 +33,5 @@ public abstract class MinecraftClientMixin {
     })
     private void cps$saveChatlogOnCrash(CallbackInfo ci) {
         ChatLog.serialize(true);
-    }
-
-    /** Replaces the vanilla chat screen with the search-enabled one */
-    @WrapOperation(method = "openChatScreen", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/MinecraftClient;setScreen(Lnet/minecraft/client/gui/screen/Screen;)V", ordinal = 1))
-    private void cps$overrideChatScreen(MinecraftClient mc, Screen vanillaChatScreen, Operation<Void> setScreenMethod, String initialText) {
-        if(ChatPatches.config.chatSearchScreen)
-            setScreenMethod.call(mc, new ChatSearchScreen(initialText));
-
-        else
-            setScreenMethod.call(mc, vanillaChatScreen);
     }
 }